For #21437 - Relocated Home-related settings to its dedicated sub screen
