+++ app/src/main/java/org/mozilla/fenix/settings/CustomizationFragment.kt
@@ -9,7 +9,6 @@ import android.os.Build
 import android.os.Build.VERSION.SDK_INT
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatDelegate
-import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
 import androidx.preference.SwitchPreference
 import org.mozilla.fenix.FeatureFlags
@@ -51,7 +50,6 @@ class CustomizationFragment : PreferenceFragmentCompat() {
         bindAutoBatteryTheme()
         setupRadioGroups()
         setupToolbarCategory()
-        setupHomeCategory()
         setupGesturesCategory()
     }
 
@@ -140,37 +138,6 @@ class CustomizationFragment : PreferenceFragmentCompat() {
         addToRadioGroup(topPreference, bottomPreference)
     }
 
-    private fun setupHomeCategory() {
-        requirePreference<SwitchPreference>(R.string.pref_key_enable_top_frecent_sites).apply {
-            isChecked = context.settings().showTopFrecentSites
-            onPreferenceChangeListener = CustomizeHomeMetricsUpdater()
-        }
-
-        requirePreference<SwitchPreference>(R.string.pref_key_recent_tabs).apply {
-            isVisible = FeatureFlags.showRecentTabsFeature
-            isChecked = context.settings().showRecentTabsFeature
-            onPreferenceChangeListener = CustomizeHomeMetricsUpdater()
-        }
-
-        requirePreference<SwitchPreference>(R.string.pref_key_recent_bookmarks).apply {
-            isVisible = FeatureFlags.recentBookmarksFeature
-            isChecked = context.settings().showRecentBookmarksFeature
-            onPreferenceChangeListener = CustomizeHomeMetricsUpdater()
-        }
-
-        requirePreference<SwitchPreference>(R.string.pref_key_pocket_homescreen_recommendations).apply {
-            isVisible = FeatureFlags.isPocketRecommendationsFeatureEnabled(context)
-            isChecked = context.settings().showPocketRecommendationsFeature
-            onPreferenceChangeListener = CustomizeHomeMetricsUpdater()
-        }
-
-        requirePreference<SwitchPreference>(R.string.pref_key_history_metadata_feature).apply {
-            isVisible = FeatureFlags.historyMetadataUIFeature
-            isChecked = context.settings().historyMetadataUIFeature
-            onPreferenceChangeListener = CustomizeHomeMetricsUpdater()
-        }
-    }
-
     private fun setupGesturesCategory() {
         requirePreference<SwitchPreference>(R.string.pref_key_website_pull_to_refresh).apply {
             isVisible = FeatureFlags.pullToRefreshEnabled
@@ -184,24 +151,6 @@ class CustomizationFragment : PreferenceFragmentCompat() {
         requirePreference<SwitchPreference>(R.string.pref_key_swipe_toolbar_switch_tabs).apply {
             isChecked = context.settings().isSwipeToolbarToSwitchTabsEnabled
             onPreferenceChangeListener = SharedPreferenceUpdater()
-        }
-    }
-
-    class CustomizeHomeMetricsUpdater : SharedPreferenceUpdater() {
-        override fun onPreferenceChange(preference: Preference, newValue: Any?): Boolean {
-            try {
-                val context = preference.context
-                context.components.analytics.metrics.track(
-                    Event.CustomizeHomePreferenceToggled(
-                        preference.key,
-                        newValue as Boolean,
-                        context
-                    )
-                )
-            } catch (e: IllegalArgumentException) {
-                // The event is not tracked
-            }
-            return super.onPreferenceChange(preference, newValue)
         }
     }
 }
