+++ app/src/main/res/values/preference_keys.xml
@@ -258,14 +258,15 @@
     <string name="pref_key_tab_view_list_do_not_use" translatable="false">pref_key_tab_view_list</string>
     <string name="pref_key_tab_view_grid" translatable="false">pref_key_tab_view_grid</string>
     <string name="pref_key_tabs" translatable="false">pref_key_tabs</string>
+    <string name="pref_key_home" translatable="false">pref_key_home</string>
     <string name="pref_key_close_tabs_manually" translatable="false">pref_key_close_tabs_manually</string>
     <string name="pref_key_close_tabs_after_one_day" translatable="false">pref_key_close_tabs_after_one_day</string>
     <string name="pref_key_close_tabs_after_one_week" translatable="false">pref_key_close_tabs_after_one_week</string>
     <string name="pref_key_close_tabs_after_one_month" translatable="false">pref_key_close_tabs_after_one_month</string>
     <string name="pref_key_allow_third_party_root_certs" translatable="false">pref_key_allow_third_party_cert_roots</string>
-    <string name="pref_key_start_on_home_after_four_hours" translatable="false">pref_key_start_on_home_after_four_hours</string>
-    <string name="pref_key_start_on_home_always" translatable="false">pref_key_start_on_home_always</string>
-    <string name="pref_key_start_on_home_never" translatable="false">pref_key_start_on_home_never</string>
+    <string name="pref_key_opening_screen_homepage" translatable="false">pref_key_opening_screen_homepage</string>
+    <string name="pref_key_opening_screen_last_tab" translatable="false">pref_key_opening_screen_last_tab</string>
+    <string name="pref_key_opening_screen_after_four_hours_of_inactivity" translatable="false">pref_key_opening_screen_after_four_hours_of_inactivity</string>
     <string name="pref_key_start_on_home_category" translatable="false">pref_key_start_on_home_category</string>
     <string name="pref_key_camera_permissions_needed" translatable="false">pref_key_camera_permissions_needed</string>
     <string name="pref_key_inactive_tabs_category" translatable="false">pref_key_inactive_tabs_category</string>
