+++ app/src/main/res/xml/home_preferences.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- This Source Code Form is subject to the terms of the Mozilla Public
+   - License, v. 2.0. If a copy of the MPL was not distributed with this
+   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
+<androidx.preference.PreferenceScreen
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <androidx.preference.SwitchPreference
+        android:key="@string/pref_key_enable_top_frecent_sites"
+        android:title="@string/top_sites_toggle_top_frecent_sites_2" />
+
+    <androidx.preference.SwitchPreference
+        android:key="@string/pref_key_recent_tabs"
+        android:title="@string/customize_toggle_jump_back_in"
+        app:isPreferenceVisible="false" />
+
+    <androidx.preference.SwitchPreference
+        android:key="@string/pref_key_recent_bookmarks"
+        android:title="@string/customize_toggle_recent_bookmarks"
+        app:isPreferenceVisible="false" />
+
+    <androidx.preference.SwitchPreference
+        android:key="@string/pref_key_history_metadata_feature"
+        android:title="@string/customize_toggle_recent_searches"
+        app:isPreferenceVisible="false" />
+
+    <androidx.preference.SwitchPreference
+        android:key="@string/pref_key_pocket_homescreen_recommendations"
+        android:title="@string/customize_toggle_pocket"
+        app:isPreferenceVisible="false" />
+
+    <androidx.preference.PreferenceCategory
+        android:layout="@layout/preference_cat_style"
+        android:title="@string/preferences_opening_screen"
+        app:allowDividerAbove="true"
+        app:isPreferenceVisible="false"
+        android:key="pref_key_start_on_home_category"
+        app:iconSpaceReserved="false">
+
+        <org.mozilla.fenix.settings.RadioButtonPreference
+            android:defaultValue="true"
+            android:key="@string/pref_key_opening_screen_homepage"
+            android:title="@string/opening_screen_homepage" />
+
+        <org.mozilla.fenix.settings.RadioButtonPreference
+            android:defaultValue="false"
+            android:key="@string/pref_key_opening_screen_last_tab"
+            android:title="@string/opening_screen_last_tab" />
+
+        <org.mozilla.fenix.settings.RadioButtonPreference
+            android:defaultValue="false"
+            android:key="@string/pref_key_opening_screen_after_four_hours_of_inactivity"
+            android:title="@string/opening_screen_after_four_hours_of_inactivity" />
+
+    </androidx.preference.PreferenceCategory>
+</androidx.preference.PreferenceScreen>
