+++ app/src/main/res/xml/preferences.xml
@@ -61,6 +61,11 @@
             android:title="@string/preferences_tabs" />
 
         <androidx.preference.Preference
+            app:iconSpaceReserved="false"
+            android:key="@string/pref_key_home"
+            android:title="@string/preferences_home_2" />
+
+        <androidx.preference.Preference
             android:key="@string/pref_key_customize"
             app:iconSpaceReserved="false"
             android:title="@string/preferences_customize" />
