+++ app/src/main/res/xml/preferences_default_browser_experiment.xml
@@ -45,6 +45,11 @@
             app:iconSpaceReserved="false" />
 
         <androidx.preference.Preference
+            app:iconSpaceReserved="false"
+            android:key="@string/pref_key_home"
+            android:title="@string/preferences_home_2" />
+
+        <androidx.preference.Preference
             android:key="@string/pref_key_customize"
             android:title="@string/preferences_customize"
             app:iconSpaceReserved="false" />
